# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FastapMenu - Digital Restaurant Dining Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants/{restaurantId}/branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List branches
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create branch
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /restaurants/{restaurantId}/branches/{branchId}:
    put:
      operationId: updateBranch
      tags: [branches]
      summary: Update branch
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
        - name: branchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      summary: Delete branch
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
        - name: branchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── TABLES ──────────────────────────────────────────────────────────────────
...
  /restaurants/{restaurantId}/categories:
    get:
      operationId: listCategories
      tags: [menus]
      summary: List menu categories
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [menus]
      summary: Create category
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /restaurants/{restaurantId}/categories/{categoryId}:
    put:
      operationId: updateCategory
      tags: [menus]
      summary: Update category
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [menus]
      summary: Delete category
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── MENU ITEMS ──────────────────────────────────────────────────────────────
  /restaurants/{restaurantId}/items:
    get:
      operationId: listItems
      tags: [items]
      summary: List menu items
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
        - name: categoryId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createItem
      tags: [items]
      summary: Create menu item
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /restaurants/{restaurantId}/items/{itemId}:
    get:
      operationId: getItem
      tags: [items]
      summary: Get menu item
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    put:
      operationId: updateItem
      tags: [items]
      summary: Update menu item
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteItem
      tags: [items]
      summary: Delete menu item
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── QR CODES ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ─── AUTH ─────────────────────────────────────────────────────────────────
...
    Branch:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, restaurantId, name, isActive, createdAt]
    BranchInput:
      type: object
      properties:
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
      required: [name]

    # ─── TABLE ────────────────────────────────────────────────────────────────
...
    Category:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        name:
          type: string
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        isAvailable:
          type: boolean
        availableFrom:
          type: ["string", "null"]
        availableTo:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, restaurantId, name, sortOrder, isAvailable, createdAt]
    CategoryInput:
      type: object
      properties:
        name:
          type: string
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        isAvailable:
          type: boolean
        availableFrom:
          type: ["string", "null"]
        availableTo:
          type: ["string", "null"]
      required: [name]

    # ─── MENU ITEM ────────────────────────────────────────────────────────────
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        categoryId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        discountedPrice:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        ingredients:
          type: ["string", "null"]
        allergens:
          type: ["string", "null"]
        calories:
          type: ["integer", "null"]
        prepTime:
          type: ["integer", "null"]
        spiceLevel:
          type: integer
        dietaryTags:
          type: array
          items:
            type: string
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ItemVariant"
        addons:
          type: array
          items:
            $ref: "#/components/schemas/ItemAddon"
        createdAt:
          type: string
          format: date-time
      required: [id, restaurantId, categoryId, name, price, spiceLevel, dietaryTags, isAvailable, isFeatured, sortOrder, variants, addons, createdAt]
    ItemVariant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        priceDiff:
          type: number
      required: [id, name, priceDiff]
    ItemAddon:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
      required: [id, name, price]
    MenuItemInput:
      type: object
      properties:
        categoryId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        discountedPrice:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        ingredients:
          type: ["string", "null"]
        allergens:
          type: ["string", "null"]
        calories:
          type: ["integer", "null"]
        prepTime:
          type: ["integer", "null"]
        spiceLevel:
          type: integer
        dietaryTags:
          type: array
          items:
            type: string
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
        variants:
          type: array
          items:
            $ref: "#/components/schemas/VariantInput"
        addons:
          type: array
          items:
            $ref: "#/components/schemas/AddonInput"
      required: [categoryId, name, price]
    VariantInput:
      type: object
      properties:
        name:
          type: string
        priceDiff:
          type: number
      required: [name, priceDiff]
    AddonInput:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
      required: [name, price]

    # ─── QR CODE ──────────────────────────────────────────────────────────────
...